requirements.txt
BeautifulSoup==3.2.1
PIL==1.1.7
amqp==1.0.13
anyjson==0.3.3
beautifulsoup4==4.1.3
celery==3.0.21
dateutils==0.6.6
django-celery==3.0.17
django-grappelli==2.4.5
flower==0.5.2
ipython==0.13.1
librabbitmq==1.0.1
matplotlib==1.1.1
pytz==2013b
redis==2.7.6
reportlab==2.5
